seam: keep the read meter's clippy sites level with the base
The gate compares (file, lint) site sets, and the meter added seven the
base does not have:

* five `private_interfaces`, from three entry points visible to the crate
  whose `SeamReadSeam` / `SeamReadTicket` are private to this module. Only
  this module's own seam and tests call them, so they are private now.
* one redundant conversion on a Vulkan handle, which `as_raw` already
  answers as a `u64` on every target this device builds for.
* one binding a test never read.

None of the three changes a decision.

+++ crates/reims-vgpu/src/backend/vulkan/engine/mod.rs
@@ -3072,8 +3072,8 @@ pub fn resident_read_witness(identity: &TargetIdentity) -> Option<ResidentReadWi
     let slot = guard.pools.registry_get(identity)?;
     Some(ResidentReadWitness {
         key_generation: identity.generation(),
-        image: slot.image.as_raw() as u64,
-        view: slot.view.as_raw() as u64,
+        image: slot.image.as_raw(),
+        view: slot.view.as_raw(),
         content_serial: (slot.content_serial != 0).then_some(slot.content_serial),
         content_epoch: slot.content_epoch,
         content_ready: slot.content_ready,
